requirements.txt
fastapi
uvicorn
pyautogen
llama-index
llama-index-embeddings-huggingface  # avoid unnecessary packages
